Add loadUsersPerMonth in ReportsService
