+++ src/Services/ReportsService.js
@@ -12,6 +12,36 @@ export const loadDisabledUsersForType = type => {
     .then(res => res.json())
 }
 
+export const loadUsersPerMonth = (startDate, endDate) => {
+  const usersData = {
+    '2017-05': {
+      users: 1,
+      premiumUsers: 0
+    },
+    '2017-06': {
+      users: 3,
+      premiumUsers: 1
+    },
+    '2017-07': {
+      users: 5,
+      premiumUsers: 2
+    },
+    '2017-08': {
+      users: 4,
+      premiumUsers: 3
+    },
+    '2017-09': {
+      users: 10,
+      premiumUsers: 7
+    }
+  }
+  return Promise.resolve(usersData)
+  // TODO: Uncomment me
+  // return fetch(`${baseUrl}/analytics/users?startDate=${startDate}&endDate=${endDate}`)
+  //   .then(handleErrors)
+  //   .then(res => res.json())
+}
+
 function handleErrors (response) {
   if (!response.ok) {
     console.log('Error:')
